Tweaks
Fixed a bug where thatch dropped silver blocks, made it so you need 8
wheat and a container to make a grain container, but the flour now
produces 3 dough when mixed with water. This is to reduce the number of
times you need to fill up a bucket and basket during the whole process,
condensing things down a bit.