Hateful Bandits
In addition to tweaking the ore process so it usually will drop the
produced item if your inventory is full, I messed with bandits. They now
torch any building they break the door of, which creates a custom fire
type I implemented which drops ash as it burns. I really like the effect
since now torched buildings are obviously so, rather than odd stone
platforms or whatever, but it might not be to everyone's liking.