Grapevines
Added grapevines that grow and provide grapes... If they are past the
first stage of ripeness, when you bust the vine it will revert back to
the first stage. If it's on the last stage when you bust it, you get
grapes. Vines can be harvested via shears and planted on dirt/grass or
other grapevines.