Weapon Changes
Removed speed buff when charging weapons, so now you're slow when
swinging again, but at least all the bugs with that are gone. Javelins
now stack properly, no longer hit you in the back when walking and
throwing weakly. Fixed bug where dropping javelins was broken (dropped
the stack that's in your inventory, as opposed to a new stack).